Console inventory UI with item usage
I: View inventory contents
0-9: Select an item and see more info about it
R: Add up to 5 random items to the inventory. For testing purposes
U: After selecting an item with 0-9 you can press U to use it if it has a use listed
P: List the players stats. You can use this before and after using a potion to see it's effects.