Initial level generation
Level generation code. Final result is a list of rooms in Level, with each Room having a list of connected rooms. A few checks are performed during level gen to make sure the level is valid:
- Floodfill done from the start room. If the end room isn't reached level generation will fail. Any rooms that aren't connected to the start room are pruned, since there isn't any way to reach them.
- Each room is checked to have >= 2 connections. If they don't level generation will fail.

Level gen failures are rare. The Level constructor will try up to 10 times to generate a level to counter failures.