New level generation technique
The goals of this technique are to have more consistent and interesting level generation, and to have less seeds that need to be discarded because they generate invalid levels. So far it seems to be hitting those goals aside for some bugs that need to be fixed. This technique works by generating several primary paths which go directly from the start room to the end room with some rooms in between. Next there's a random chance that secondary paths will branch off from each room on the primary path. At the moment these just connect two primary paths or connect two rooms on one primary path. One future improvement is to make more complex multi-room branches. The final step is to prune some connections. This stops the level from being overly connected and often breaks up primary paths so you can't go straight from the start room to the end room along a primary path.