Make a deep copy of level when moving it between state buffers
Since level is class we were passing around a reference to the same class instance. This can cause issues with the logic and render threads if one were reading data while the other was editing it. Particularly when one is iterating lists while the other is adding/removing items from it.