Fix possible bug in item pickup code
It was removing items from the list while iterating it. This isn't good practice and can cause crashes. It now makes a copy of the list and iterates that while removing items from the original list. This isn't very performant but we have so few items that it doesn't matter.