Level cloning improvements
This fixes an oversight in Level.Clone() that meant that Rooms in cloned levels still referenced rooms in the original level instead of reference clones. So far this seems to have fixed a bug where the wrong room would be renderered. I tested it on seeds I noticed the issue on previously and it was no longer there. The bug was inconsistent so it's hard to say with complete certainty that it's fixed, will need to keep a look out while testing the game.