Add special door to end rooms. Must walk through to get to next level
Previously you'd get teleported to the next level if you were anywhere in the end room and had all keys. This is a bit better since you don't just randomly get teleported. The texture used is a placeholder. Don't have time at the moment to make/find a better one.